refactor: isolate render
