+++ apps/client/src/features/rundown/block-block/BlockBlock.tsx
@@ -1,14 +1,13 @@
 import { useRef } from 'react';
-import { IconButton } from '@chakra-ui/react';
 import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
 import { IoReorderTwo } from '@react-icons/all-files/io5/IoReorderTwo';
-import { IoTrash } from '@react-icons/all-files/io5/IoTrash';
 import { OntimeBlock } from 'ontime-types';
 
-import { AppMode, useAppMode } from '../../../common/stores/appModeStore';
 import { cx } from '../../../common/utils/styleUtils';
 import EditableBlockTitle from '../common/EditableBlockTitle';
+
+import BlockDelete from './BlockDelete';
 
 import style from './BlockBlock.module.scss';
 
@@ -21,7 +20,6 @@ interface BlockBlockProps {
 export default function BlockBlock(props: BlockBlockProps) {
   const { data, hasCursor, onDelete } = props;
 
-  const mode = useAppMode((state) => state.mode);
   const handleRef = useRef<null | HTMLSpanElement>(null);
 
   const {
@@ -41,7 +39,6 @@ export default function BlockBlock(props: BlockBlockProps) {
   };
 
   const blockClasses = cx([style.block, hasCursor ? style.hasCursor : null]);
-  const isRunMode = mode === AppMode.Run;
 
   return (
     <div className={blockClasses} ref={setNodeRef} style={dragStyle}>
@@ -49,15 +46,7 @@ export default function BlockBlock(props: BlockBlockProps) {
         <IoReorderTwo />
       </span>
       <EditableBlockTitle title={data.title} eventId={data.id} placeholder='Block title' />
-      <IconButton
-        aria-label='Delete'
-        size='sm'
-        icon={<IoTrash />}
-        variant='ontime-subtle'
-        color='#FA5656'
-        onClick={onDelete}
-        isDisabled={isRunMode}
-      />
+      <BlockDelete onDelete={onDelete} />
     </div>
   );
 }
